Only download binaries that are missing during compile

+++ compile-windows.py
@@ -11,42 +11,46 @@ except ModuleNotFoundError:
     os.system('pip install pyunpack patool')
     from pyunpack import Archive
 
-print('Downloading ffmpeg')
-with open('ffmpeg.7z', 'wb+') as f:
-    f.write(requests.get('https://www.gyan.dev/ffmpeg/builds/ffmpeg-git-essentials.7z').content)
+if os.path.isfile('ffmpeg.exe') == False:
+    print('Downloading ffmpeg')
+    with open('ffmpeg.7z', 'wb+') as f:
+        f.write(requests.get('https://www.gyan.dev/ffmpeg/builds/ffmpeg-git-essentials.7z').content)
 
-print('Downloading ImageMagick')
-with open('ImageMagick.zip', 'wb+') as f:
-    f.write(requests.get('https://imagemagick.org/archive/binaries/ImageMagick-7.1.0-portable-Q16-x64.zip').content)
+    print('Extracting ffmpeg')
+    os.makedirs('ffmpeg', exist_ok=True)
+    Archive('ffmpeg.7z').extractall("ffmpeg")
+    for i in os.listdir('ffmpeg'):
+        i_path = os.path.join('ffmpeg', i)
+        for j in os.listdir(i_path):
+            j_path = os.path.join(i_path, j)
+            os.rename(j_path, os.path.join('ffmpeg', j))
 
-print('Downloading zip')
-with open('zip.zip', 'wb+') as f:
-    f.write(requests.get('https://sourceforge.net/projects/gnuwin32/files/zip/3.0/zip-3.0-bin.zip').content)
+if os.path.isfile('magick.exe') == False:
+    print('Downloading ImageMagick')
+    with open('ImageMagick.zip', 'wb+') as f:
+        f.write(requests.get('https://imagemagick.org/archive/binaries/ImageMagick-7.1.0-portable-Q16-x64.zip').content)
 
-print('Downloading bzip2')
-with open('bzip2.zip', 'wb+') as f:
-    f.write(requests.get('https://sourceforge.net/projects/gnuwin32/files/bzip2/1.0.5/bzip2-1.0.5-bin.zip').content)
+    print('Extracting ImageMagick')
+    os.makedirs('ImageMagick', exist_ok=True)
+    Archive('ImageMagick.zip').extractall("ImageMagick")
 
-print('Extracting ffmpeg')
-os.makedirs('ffmpeg', exist_ok=True)
-Archive('ffmpeg.7z').extractall("ffmpeg")
-for i in os.listdir('ffmpeg'):
-    i_path = os.path.join('ffmpeg', i)
-    for j in os.listdir(i_path):
-        j_path = os.path.join(i_path, j)
-        os.rename(j_path, os.path.join('ffmpeg', j))
+if os.path.isfile('zip.exe') == False:
+    print('Downloading zip')
+    with open('zip.zip', 'wb+') as f:
+        f.write(requests.get('https://sourceforge.net/projects/gnuwin32/files/zip/3.0/zip-3.0-bin.zip').content)
 
-print('Extracting ImageMagick')
-os.makedirs('ImageMagick', exist_ok=True)
-Archive('ImageMagick.zip').extractall("ImageMagick")
+    print('Extracting zip')
+    os.makedirs('zip', exist_ok=True)
+    Archive('zip.zip').extractall("zip")
 
-print('Extracting zip')
-os.makedirs('zip', exist_ok=True)
-Archive('zip.zip').extractall("zip")
+if os.path.isfile('bzip2.exe') == False:
+    print('Downloading bzip2')
+    with open('bzip2.zip', 'wb+') as f:
+        f.write(requests.get('https://sourceforge.net/projects/gnuwin32/files/bzip2/1.0.5/bzip2-1.0.5-bin.zip').content)
 
-print('Extracting bzip2')
-os.makedirs('bzip2', exist_ok=True)
-Archive('bzip2.zip').extractall("bzip2")
+    print('Extracting bzip2')
+    os.makedirs('bzip2', exist_ok=True)
+    Archive('bzip2.zip').extractall("bzip2")
 
 print('Creating exe')
 os.system('pip install pyinstaller')
